+++ ThaumAge/Assets/Scrpits/Bean/Common/BaseBean.cs
@@ -5,6 +5,6 @@ using System;
 [Serializable]
 public class BaseBean 
 {
-    public long id;//id
+    public int id;//id
     public int valid;//是否有效
 }
